pyproject.toml
[project]
name = 'astroimtools'
description = 'Astronomical Image Tools'
readme = 'README.rst'
license = 'BSD-3-Clause'
license-files = ['LICENSE.rst']
authors = [
    {name = 'Astroimtools Developers', email = 'help@stsci.edu'},
]
keywords = [
    'astronomy',
    'astrophysics',
    'images',
    'cutouts',
    'statistics',
    'filtering',
]
classifiers = [
    'Intended Audience :: Science/Research',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Topic :: Scientific/Engineering :: Astronomy',
]
dynamic = ['version']
requires-python = '>=3.9'
dependencies = [
    'numpy>=1.22',
    'astropy>=5.0',
    'scipy>=1.7.2',
]

[project.urls]
Homepage = 'https://github.com/spacetelescope/astroimtools'
Documentation = 'https://astroimtools.readthedocs.io/en/stable/'

[project.scripts]
imarith = "astroimtools.scripts.imarith:main"
imstats = "astroimtools.scripts.imstats:main"

[project.optional-dependencies]
all = [
    'matplotlib>=3.5',
    'photutils>=1.5',
    'reproject>=0.9',
]
test = [
    'pytest-astropy>=0.10',
]
docs = [
    'astroimtools[all]',
    'ipython',
    'nbsphinx',
    'sphinx',
    'sphinx-astropy>=1.6',
    'tomli; python_version < "3.11"',
]

[build-system]
requires = [
    'setuptools>=61.2',
    'setuptools_scm>=6.2',
]
build-backend = 'setuptools.build_meta'

[tool.setuptools_scm]
write_to = 'astroimtools/version.py'

[tool.setuptools]
zip-safe = false
include-package-data = false

[tool.setuptools.packages.find]
namespaces = false

[tool.setuptools.package-data]
'astroimtools' = [
    'data/*',
]

[tool.pytest.ini_options]
minversion = 7.0
testpaths = [
    'astroimtools',
    'docs',
]
norecursedirs = [
    'docs/_build',
    'extern',
]
astropy_header = true
doctest_plus = 'enabled'
text_file_format = 'rst'
addopts = [
    '-ra',
    '--color=yes',
    '--doctest-rst',
    '--strict-config',
    '--strict-markers',
]
log_cli_level = 'INFO'
xfail_strict = true
remote_data_strict = true
filterwarnings = [
    'error',  # turn warnings into exceptions
]

[tool.coverage.run]
omit = [
    'astroimtools/_astropy_init*',
    'astroimtools/conftest.py',
    'astroimtools/*setup_package*',
    'astroimtools/tests/*',
    'astroimtools/*/tests/*',
    'astroimtools/extern/*',
    'astroimtools/version*',
    '*/astroimtools/_astropy_init*',
    '*/astroimtools/conftest.py',
    '*/astroimtools/*setup_package*',
    '*/astroimtools/tests/*',
    '*/astroimtools/*/tests/*',
    '*/astroimtools/extern/*',
    '*/astroimtools/version*',
]

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'except ImportError',
    'raise AssertionError',
    'raise NotImplementedError',
    'def main\\(.*\\):',
    'pragma: py{ignore_python_version}',
    'def _ipython_key_completions_',
]

[tool.build-sphinx]
github_project = 'spacetelescope/astroimtools'

[tool.isort]
skip_glob = [
    'astroimtools/*__init__.py*',
]
known_first_party = [
    'astroimtools',
    'extension_helpers',
]
use_parentheses = true

[tool.black]
force-exclude = """
(
  .*
)
"""

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.repo-review]
ignore = [
    'MY',  # ignore MyPy
    'PC110',  # ignore using black or ruff-format in pre-commit
    'PC111',  # ignore using blacken-docs in pre-commit
    'PC140',  # ignore using mypy in pre-commit
    'PC180',  # ignore using prettier in pre-commit
    'PC901',  # ignore using custom pre-commit update message
    'PY005',  # ignore having a tests/ folder
]

[tool.docformatter]
    wrap-summaries = 72
    pre-summary-newline = true
    make-summary-multi-line = true

[tool.numpydoc_validation]
checks = [
    'all',   # report on all checks, except the below
    'ES01',  # missing extended summary
    'EX01',  # missing "Examples"
    'GL08',  # missing docstring
    'RT01',  # do not require return type for lazy properties
    'RT02',  # only type in "Returns" section (no name)
    'SA01',  # missing "See Also"
    'SA04',  # missing "See Also" description
    'SS06',  # single-line summary
]

# don't report on objects that match any of these regex;
# remember to use single quotes for regex in TOML
exclude = [
    '\._.*',  # private functions/methods
    '^test_*',  # test code
    '^conftest.*$',  # pytest configuration
]

[tool.ruff.lint]
select = [
    'B', 'D', 'E', 'F', 'FLY', 'I', 'INT', 'N', 'NPY', 'PERF', 'PL', 'Q',
    'RSE', 'SIM', 'UP', 'W'
]
ignore = [
    'E501',
    'B028',
    'B905',  # revisit
    'D100',
    'D101',
    'D102',
    'D103',
    'D105',
    'D107',
    'D200',
    'D205',
    'D212',
    'D401',
    'D404',
    'I001',
    'PLC0415',  # import-outside-top-level
    'PLR0912',
    'PLR0913',
    'PLR0915',
    'PLR2004',
    'PLW1641',  # eq-without-hash
    'PLW2901',
    'Q000',
    'SIM910',
    'UP038',
]

[tool.ruff.lint.per-file-ignores]
'__init__.py' = ['D104', 'I']
'docs/conf.py' = ['ERA001', 'INP001', 'TRY400']

[tool.ruff.lint.pydocstyle]
convention = 'numpy'
